Edited Files: - RiotManager.java - RiotModel.java - CollegeModel.java
Changes:
Added severity conditions for riots, terrible conditions result in SevereRiot, one terrible condition leads to RowdyBehaviour, and poor Faculty Happiness and pay leads to FacultyRiot. Added a location for RiotModel that returns a String of the affected building.

To Do:
Have RiotModel.java have a reference to a BuildingModel object. When a riot occurs, chance of the building being damaged or de-leveled.